test3

+++ deploy/scripts/init_ftr.sh
@@ -3,6 +3,9 @@ cd /var/www/html/ftr.mixcart.ru/
 export COMPOSER_HOME="$HOME/.composer";
 composer update
 php yii migrate --interactive=0
+if [ $? -gt 0 ]; then
+    exit 1
+fi
 
 sudo rm -rf /var/www/html/ftr.mixcart.ru/frontend/web/assets/*
 sudo rm -rf /var/www/html/ftr.mixcart.ru/backend/web/assets/*
